requirements.txt
certifi==2025.6.15
charset-normalizer==3.4.2
contourpy==1.3.2
cycler==0.12.1
docker==7.1.0
fonttools==4.58.4
idna==3.10
kafka-python==2.2.12
kiwisolver==1.4.8
matplotlib==3.10.3
numpy==2.3.0
packaging==25.0
pandas==2.3.0
pillow==11.2.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.4
six==1.17.0
style==1.1.0
tzdata==2025.2
update==0.0.1
urllib3==2.5.0
